+ air humidity

+++ src/Plant.java
@@ -1,8 +1,8 @@
 public interface Plant {
-    double minSoilMoisture = 0; //минимальная влажность почвы, при которой растение не засохнет
-    double maxSoilMoisture = 0; //максимальная влажность почвы, при которой растение не загниет
-    double minAirHumidity = 0; //максимальная влажность воздуха
-    double maxAirHumidity = 0; //минимальная влажность воздуха
+    float minSoilMoisture = 0; //минимальная влажность почвы, при которой растение не засохнет
+    float maxSoilMoisture = 0; //максимальная влажность почвы, при которой растение не загниет
+    float minAirHumidity = 0; //максимальная влажность воздуха
+    float maxAirHumidity = 0; //минимальная влажность воздуха
 
 
 }
